Create Select_Variant.sh
This script selects variant in genes of interest from the VCF files check for filters and annotates them with VEP. It will generate a intersected bed file, a filtered and VEP annotated VCF file. and a consolidated standard out with varients in genes of interest present in all samples and their coordinates which can then be piped to a text file.